Added rudimentary HTTP streaming support.
Currently this is not optimized at all. It takes over a thousand HTTP requests (although libcurl does keep the connection alive at least) to render one page of a 35MB test PDF.

On the plus side: There is a cache, so blocks that have been downloaded will not be downloaded again. Also, while testing, no change to pdftoppm was necessary, since the PDFDoc takes care of parsing the filename as a URL.

Note that no changes were necessary outside of poppler itself, meaning the API did not change at all. It just accepts filenames starting with http:// now.